aitool-api/requirements.txt
blinker==1.9.0
Brotli==1.0.9
certifi==2025.1.31
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==3.0.10
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.0
MarkupSafe==3.0.2
PySocks==1.7.1
python-docx==1.1.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
win-inet-pton==1.1.0
